refactor(tests): extract mock utilities into centralized mocks.js
- Extracted createMockElement, createMockChrome, and createMockDocument from tests/popup.test.js into tests/mocks.js.
- Refactored tests/popup.test.js and tests/security.test.js to use the new mock utilities.
- Simplified setupPopupSandbox and improved mock robustess.
